Make the Result class act like an array by extending Iterator.
Renaming getRow to getNextValue turns it into the function that
generates the values the iterator lists. Also remove the variable
that was going to be a local cache of the rows, which is no longer
needed, and ensure that if there are no more rows we return null
and don't error. Move the call to get the fields' types after the
null check for efficiency.